+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 # 0.0.3
 
++ Improve parser
++ Replace `Pi`, `Sigma` with `Dt`
+
 # 0.0.2
 
 + Implemented a STLC core language